Add Buyer Request Crop Functionality

+++ cropMatch-App/src/main/java/com/cropMatch/repository/CropImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CropImageRepository extends JpaRepository<CropImage, Long> {
+public interface CropImageRepository extends JpaRepository<CropImage, Integer> {
 }
